Added -forceReload flag which forces page reload on any file change (#4)
I'm developing a blog which is using fetch to lazy load html pages. So
currently it doesnt refresh whenever the lazy loaded html is changed
since origins html page won't ever be the same.

So this is a quick and dirty way of solving the problem, might also
solve some similar problems down the line.